build with Multi-threaded DLL option. use ARD.lib for 2DVisualizer

+++ 2DVisualizer/transforms.cpp
@@ -9,5 +9,5 @@ Uint8 quantize_to_uint8(double input, double max) {
     return 0;
   }
 
-  return output;
+  return (Uint8)output;
 }
